dunphy-family-vacation-debate-main/pyproject.toml
[project]
name = "family-vacation-debate"
version = "0.1.0"
description = "Multi-agent Dunphy family vacation debate simulation"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.129.0",
  "google-adk==1.11.0",
  "google-genai==1.63.0",
  "python-dotenv==1.0.1",
  "uvicorn==0.40.0",
]

[tool.uv]
package = false
